+++ src/main/java/com/alechilles/alecstamework/api/TameworkApi.java
@@ -17,6 +17,8 @@ public interface TameworkApi {
 
     InteractionExtensionApi interactionExtensions();
 
+    TraitEffectApi traitEffects();
+
     ProfileDataApi profileData();
 
     TameworkEventsApi events();